requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openai>=1.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
reportlab>=4.0.0
Pillow>=10.0.0
spacy>=3.5.0
networkx>=3.1
tqdm>=4.65.0
dateparser>=1.1.8
sqlalchemy>=2.0.0
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0
streamlit>=1.28.0
pytest>=7.4.0
alembic>=1.12.0
